fix bug: rtnContextBuff::truncate后计算缓存长度存在问题
git-svn-id: http://192.168.20.11/sequoiadb/trunk@21949 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/rtn/rtnContextBuff.cpp
@@ -91,7 +91,7 @@ namespace engine
          try
          {
             BSONObj objTemp( &_pBuff[_curOffset] ) ;
-            offset += ossAlign4( objTemp.objsize() ) ;
+            offset += ossAlign4( (UINT32)objTemp.objsize() ) ;
             ++recordNum ;
          }
          catch( std::exception &e )
